feat: complete project scenarios 15-16 and user scenario 19

+++ src/projects/services.py
@@ -1,11 +1,11 @@
 import json
 from math import ceil
 from typing import Tuple
-from fastapi import UploadFile, HTTPException, status
-from pydantic_core import ValidationError
 
+from fastapi import UploadFile, HTTPException, status
 from projects.repositories import ProjectRepo
-from projects.schemas import ProjectInDB, ProjectCreate
+from projects.schemas import ProjectInDB, ProjectCreate, ProjectWithStepsInDB
+from pydantic_core import ValidationError
 from utils import create_field_map_for_model, parse_ordering, FileService, clean_errors
 
 
@@ -14,6 +14,18 @@ class ProjectService:
 
     def __init__(self, repo: ProjectRepo):
         self._repo = repo
+
+    async def get_by_id(
+            self,
+            project_id: int,
+    ) -> ProjectWithStepsInDB:
+        project = await self._repo.get_by_id(project_id, join_steps=True)
+        if not project:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f'Project with ID {project_id} not found',
+            )
+        return ProjectWithStepsInDB.model_validate(project)
 
     async def get_all(
             self,
@@ -25,7 +37,7 @@ class ProjectService:
     ) -> Tuple[list[ProjectInDB], int, int]:
         order_column, order_direction = parse_ordering(ordering, self.field_map)
 
-        result, total = self._repo.get_all(
+        result, total = await self._repo.get_all(
             search=search,
             order_column=order_column,
             order_direction=order_direction,
@@ -79,4 +91,13 @@ class ProjectService:
         return ProjectInDB.model_validate(project)
 
     async def delete(self, project_id: int) -> None:
+        project = await self._repo.get_by_id(project_id, join_steps=False)
+        if not project:
+            raise HTTPException(
+                status_code=status.HTTP_404_NOT_FOUND,
+                detail=f'Project with ID {project_id} not found',
+            )
+
         await self._repo.delete(project_id)
+        document_path = await FileService.construct_full_path_from_relative_path(project.document_path)
+        await FileService.delete_file_from_fs(document_path)
